Model split-up
