+++ Sources/PrismCore/Providers/Figma/Figma+Configuration.swift
@@ -1,8 +1,9 @@
 //
-//  File.swift
-//  
+//  Figma+Configuration.swift
+//  Prism
 //
 //  Created by Shai Mishali on 04/10/2021.
+//  Copyright © 2021 Gett. All rights reserved.
 //
 
 import Foundation
