+++ Sources/PrismCore/Providers/Zeplin/Zeplin+Generate.swift
@@ -1,8 +1,9 @@
 //
 //  Zeplin+Generate.swift
-//  
+//  Prism
 //
 //  Created by Shai Mishali on 02/10/2021.
+//  Copyright © 2021 Gett. All rights reserved.
 //
 
 import Foundation
