+++ Sources/PrismCore/Providers/Zeplin/Zeplin+Initialize.swift
@@ -1,8 +1,9 @@
 //
-//  File.swift
-//  
+//  Zeplin+Initialize.swift
+//  Prism
 //
 //  Created by Shai Mishali on 12/10/2021.
+//  Copyright © 2021 Gett. All rights reserved.
 //
 
 import Foundation
